Fix the mess with a hyp_view.rsc that did not match its definition file any more.
The hyp_view.rsc in the main directory was a multilayer
resource of RSM at some time, but did not work anymore (RSM
refuses to load the definition file). Also, XaAES refuses to
load such files. It has been replaced now by the single-language,
english resource from the subdirectory.

Also added a README in that directory.